New func with non-free. And float in struct

+++ lab_10_04_01/headers/base.h
@@ -14,7 +14,7 @@ typedef struct
     char *surname;
     int *birthday;
     int q_marks;
-    int *marks;
+    float *marks;
 }students;
 
 int free_students();
